requirements.txt
Pillow == 10.2.0 # used for uploading the UI as images to the ePaper display
spidev == 3.6 # used for connecting to the ePaper display
evdev == 1.7.0 # read generic input events. JBL Flip 5 sends the play/pause events here.
dbus-next == 0.2.3 # Linux uses Dbus to communicate with BT devices
microdot == 2.0.5 # simple web server
gpiod == 2.1.3 # handling GPIO pins for the screen and buttons